breaking WIP, changing the RPC call logic functions
