+++ web3_proxy/src/frontend/users.rs
@@ -35,7 +35,7 @@ use ethers::abi::{AbiEncode, Error, ParamType, Token};
 use ethers::prelude::sourcemap::parse;
 use ethers::prelude::H256;
 use ethers::types::{Log, Transaction, TransactionReceipt, U256};
-use ethers::utils::hex;
+use ethers::utils::{hex, keccak256};
 use ethers::{
     prelude::{Address, EthEvent},
     types::Bytes,
@@ -588,252 +588,6 @@ pub async fn user_balance_get(
 
     // TODO: Gotta create a new table for the spend part
     Ok(Json(response).into_response())
-}
-
-/// `POST /user/balance/:tx_hash` -- Manually process a confirmed txid to update a user's balance.
-///
-/// We will subscribe to events to watch for any user deposits, but sometimes events can be missed.
-/// TODO: change this. just have a /tx/:txhash that is open to anyone. rate limit like we rate limit /login
-#[debug_handler]
-pub async fn user_balance_post(
-    Extension(app): Extension<Arc<Web3ProxyApp>>,
-    TypedHeader(Authorization(bearer)): TypedHeader<Authorization<Bearer>>,
-    Path(mut params): Path<HashMap<String, String>>,
-) -> Web3ProxyResponse {
-    // Check that the user is logged-in and authorized. We don't need a semaphore here btw
-    let (caller, _semaphore) = app.bearer_is_authorized(bearer).await?;
-
-    // Get the transaction hash, and the amount that the user wants to top up by.
-    // Let's say that for now, 1 credit is equivalent to 1 dollar (assuming any stablecoin has a 1:1 peg)
-    let tx_hash: H256 = params
-        .remove("tx_hash")
-        // TODO: map_err so this becomes a 500. routing must be bad
-        .ok_or(Web3ProxyError::BadRequest(
-            "You have not provided the tx_hash in which you paid in".to_string(),
-        ))?
-        .parse()
-        .context("unable to parse tx_hash")?;
-
-    // We don't check the trace, the transaction must be a naive, simple send transaction (for now at least...)
-    // TODO: Get the respective transaction ...
-    let db_conn = app.db_conn().context("query_user_stats needs a db")?;
-    let db_replica = app
-        .db_replica()
-        .context("query_user_stats needs a db replica")?;
-
-    // Return straight false if the tx was already added ...
-    let receipt = increase_balance_receipt::Entity::find()
-        .filter(increase_balance_receipt::Column::TxHash.eq(hex::encode(tx_hash)))
-        .one(&db_conn)
-        .await?;
-    if receipt.is_some() {
-        return Err(Web3ProxyError::BadRequest(
-            "The transaction you provided has already been accounted for!".to_string(),
-        ));
-    }
-    debug!("Receipt: {:?}", receipt);
-    // Just iterate through all logs, and add them to the transaction list if there is any
-    // Address will be hardcoded in the config
-    let authorization = Arc::new(InternalAuthorization::internal(None).unwrap());
-    let transaction_receipt: TransactionReceipt = match app
-        .balanced_rpcs
-        .best_available_rpc(&authorization, None, &[], None, None)
-        .await
-    {
-        Ok(OpenRequestResult::Handle(handle)) => {
-            // TODO: Figure out how to pass the transaction hash as a parameter ...
-            warn!(
-                "Params are: {:?}",
-                &vec![format!("0x{}", hex::encode(tx_hash))]
-            );
-            handle
-                .request(
-                    "eth_getTransactionReceipt",
-                    &vec![format!("0x{}", hex::encode(tx_hash))],
-                    Level::Trace.into(),
-                    None,
-                )
-                .await
-                // TODO: What kind of error would be here
-                .map_err(|err| Web3ProxyError::Anyhow(err.into()))
-        }
-        // TODO: Probably skip this case (?)
-        Ok(_) => {
-            // TODO: actually retry?
-            // TODO: Is this the right error message?
-            Err(Web3ProxyError::NoHandleReady)
-        }
-        Err(err) => {
-            // TODO: Just skip this part until one item responds ...
-            log::trace!(
-                "cancelled funneling transaction {} from: {:?}",
-                tx_hash,
-                err,
-            );
-            Err(err)
-        }
-    }?;
-    debug!("Tx receipt: {:?}", transaction_receipt);
-
-    // Go through all logs, this should prob capture it,
-    // At least according to this SE logs are just concatenations of the underlying types (like a struct..)
-    // https://ethereum.stackexchange.com/questions/87653/how-to-decode-log-event-of-my-transaction-log
-
-    // Make sure there is only a single log within that transaction ...
-    // I don't know how to best cover the case that there might be multiple logs inside
-
-    for log in transaction_receipt.logs {
-        warn!("Should be all from the deposit contract {:?}", log.address);
-        if format!("0x{}", hex::encode(log.address)) != app.config.deposit_contract {
-            warn!(
-                "Wrong log address: {:?} {:?}",
-                hex::encode(log.address),
-                app.config.deposit_contract
-            );
-            continue;
-        }
-
-        // Get the topics out
-        let token: Address = Address::from(log.topics.get(1).unwrap().to_owned()); // .split_off(12).into();
-        let recipient_account: Address = Address::from(log.topics.get(2).unwrap().to_owned()); // .split_off(12).into();
-
-        warn!("Check the data if we can decode it {:?}", log.data);
-        // TODO: Will this work? Depends how logs are encoded
-        // recipient_account, token,
-        let amount: U256 = match ethers::abi::decode(
-            &[
-                // ParamType::Address,
-                // ParamType::Address,
-                ParamType::Uint(256usize),
-            ],
-            &log.data,
-        ) {
-            Ok(tpl) => Ok(tpl
-                .get(0)
-                .unwrap()
-                .clone()
-                .into_uint()
-                .context("Could not decode amount")?),
-            Err(err) => Err(Web3ProxyError::BadRequest(format!(
-                "Log could not be decoded: {:?}",
-                err
-            ))),
-        }?;
-
-        warn!(
-            "Coded items are: {:?} {:?} {:?}",
-            hex::encode(recipient_account),
-            hex::encode(token),
-            amount
-        );
-        // warn!("Recipient account is: ")
-
-        warn!("Recipient address is: {:?}", recipient_account);
-        warn!("Recipient address is: {:?}", recipient_account.encode());
-
-        // First, find all users ...
-        let all_users = user::Entity::find().all(db_replica.conn()).await?;
-        warn!("All users are: {:?}", all_users);
-
-        // Encoding is inefficient, revisit later
-        let recipient = match user::Entity::find()
-            .filter(user::Column::Address.eq(&recipient_account.encode()[12..]))
-            .one(db_replica.conn())
-            .await?
-        {
-            Some(x) => Ok(x),
-            None => Err(Web3ProxyError::BadRequest(
-                "The user must have signed up first. They are currently not signed up!".to_string(),
-            )),
-        }?;
-
-        // Skip if the token is not of interest, otherwise add a balance to the contract (according to the amount)
-        // TODO: Maybe check for accepted tokens here
-        // if !app.config.accepted_deposit_tokens.contains(&hex::encode(token)) {}// Could also just hardcode all the tokens we accept in the yaml
-
-        let status_code: StatusCode;
-        let response_json: serde_json::Value;
-        // TODO: Can change with accepted currencies hashmap
-        // let target_token = "c9fcfa7e28ff320c49967f4522ebc709aa1fde7c".parse::<Address>().unwrap();
-        // debug!("Token vs target token are: {:?} {:?}", token, target_token);
-        // if token == target_token {
-        // TODO: Check for accepted tokens, and how to get them
-
-        // Increase balance by amount (whatever amount corresponds to ...)
-        // let's say 1 USD is 1_000_000 credits
-        // Basically credits are in the order of 10^6 per dollar
-        warn!("Checking amount");
-        //     .checked_div(U256::from(10).checked_pow(U256::from(12)).unwrap())
-        //     .context("Division error, this should never happen")?.as_u64();
-
-        // TODO: Make the conversion from token to credits here ...
-        // The balance can be dollars (credits are separate)
-
-        // For now, let's say that we have USDC / 6 decimal tokens
-        // so 1e6 = 1$ = Decimal(1)
-        // TODO: Let's assume that people don't buy too much at _once_
-        let amount = Decimal::from(amount.as_u128()) / Decimal::from(1e6 as u64);
-
-        // Check if the item is in the database. If it is not, then add it into the database
-        let user_balance = balance::Entity::find()
-            .filter(balance::Column::UserId.eq(recipient.id))
-            .one(&db_conn)
-            .await?;
-        debug!("User balance is: {:?}", user_balance);
-
-        let txn = db_conn.begin().await?;
-        match user_balance {
-            Some(user_balance) => {
-                let balance_plus_amount = user_balance.available_balance + amount;
-                debug!("New user balance is: {:?}", balance_plus_amount);
-                // Update the entry, adding the balance
-                let mut user_balance = user_balance.into_active_model();
-                user_balance.available_balance = sea_orm::Set(balance_plus_amount);
-                debug!("New user balance model is: {:?}", user_balance);
-                user_balance.save(&txn).await?;
-                // txn.commit().await?;
-                // user_balance
-            }
-            None => {
-                // Create the entry with the respective balance
-                let user_balance = balance::ActiveModel {
-                    available_balance: sea_orm::ActiveValue::Set(amount),
-                    // used_balance: sea_orm::ActiveValue::Set(0),
-                    user_id: sea_orm::ActiveValue::Set(recipient.id),
-                    ..Default::default()
-                };
-                debug!("New user balance model is: {:?}", user_balance);
-                user_balance.save(&txn).await?;
-                // txn.commit().await?;
-                // user_balance // .try_into_model().unwrap()
-            }
-        };
-        debug!("Setting tx_hash: {:?}", tx_hash);
-        let receipt = increase_balance_receipt::ActiveModel {
-            tx_hash: sea_orm::ActiveValue::Set(hex::encode(tx_hash)),
-            ..Default::default()
-        };
-        receipt.save(&txn).await?;
-        txn.commit().await?;
-        debug!("Submitted saving");
-
-        // Can return here
-        debug!("Returning response");
-        let response = (
-            StatusCode::CREATED,
-            Json(json!({
-                "tx_hash": tx_hash,
-                "amount": amount
-            })),
-        )
-            .into_response();
-        // Return early if the log was added
-        return Ok(response.into());
-    }
-
-    Err(Web3ProxyError::BadRequest(
-        "No such transaction was found, or token is not supported!".to_string(),
-    ))
 }
 
 /// `GET /user/keys` -- Use a bearer token to get the user's api keys and their settings.
